[Workflows] Add additional search metadata to `surfaced_in_list` event (Recidiviz/recidiviz-dashboards#6427)
Add details relating to the search that brought up a person when the `frontend.surfaced_in_list` event is emitted: the search ID value (a facility ID, facility unit ID, or the person's pseudonymized officer ID), the search field (type of search), and the tab the person was in at the time. This will be used in the new Usage dashboard to calculate the number of caseloads with "surfaceable" clients that have been searched.

GitOrigin-RevId: 16ee32e99173912ef007cf6612836ec623b68eb7