[Workflows][US_PA] Map ADM and SPC supervision levels to copy (Recidiviz/recidiviz-dashboards#6421)
In the clients page of PA workflows, clients on "limited" supervision are currently showing as "unknown" supervision level; The backend was updated to use specific limited supervision sub-categories for the record supervisionLevel in Recidiviz/recidiviz-dataRecidiviz/recidiviz-dashboards#33374. We just need to map "ADM" -> "Administrative" and "SPC" -> "Special Circumstances" on the front end.

This PR adds a set of filter options for the supervision level field for PA clients.

GitOrigin-RevId: 4850ffcee392caca4790674fbc34762f4db459db