[Workflows][Actionability] Enable drag/drop to reorder opportunity tabs (Recidiviz/recidiviz-dashboards#6323)
Display all Eligibility Status tabs on Opportunity pages, including empty ones, and enable the user to rearrange them via drag-and-drop, saving the user's preferred tab order per opportunity to Firestore. Meanwhile, when a particular tab group is not sortable, we continue to display them in a fixed order and the user cannot rearrange them. Currently the only sortable tab group is "Eligibility Status".

To do so, we introduce the dnd-kit library. Specifically we are using the sortable preset for an ordered list of tabs on an Opportunity page.

GitOrigin-RevId: 989a5454134c3b659534f14d22de0fdcd5beb62f