[Workflows][US_AZ] Expand timeline on JII profile page (Recidiviz/recidiviz-dashboards#6520)
Move the timeline to the top of the JII page and expand it to the full width of the page, for users with the new feature variant fullWidthTimeline. This change is tagged for Arizona because it'll be part of the rollout of new features to that state along with the new Workflows, and will prepare for potential future changes highlighting more dates on the timeline.

GitOrigin-RevId: 26573b66f4ae3222963f29990e9f73d824a2259a