[Insights][Zero Grants] Update UI and hover behavior for officer opp list on supervisor page (Recidiviz/recidiviz-dashboards#6554)
First of a few UI tweaks to some supervisor homepage components for the zero grant UI designs. Confirmed these changes should apply to the supervisor homepage across the board, so we don't need to gate them behind the zero grants FV.

This PR updates the workflows officer list UI on the supervisor page:

Remove the arrow hover state and add a background color change instead.
Tweaks the text color and adds an underline to the officer name on hover.
Makes the hover target the entire list item, not just the officer name.

GitOrigin-RevId: 04d50bab803c330ee2dc8d2c05fdd660e71be14b