[Insights][Zero Grants] Add zero grants pill to officer view and opportunity summaries (Recidiviz/recidiviz-dashboards#6681)
Adds the zero grants pill to the officer view and handles other requested design tweaks such as adding a larger hover target to the opportunity type summaries and a background color change on hover.

The larger hover target is achieved by wrapping the entire opportunity type summary in a Link
The zero grants pill is added by optionally passing in the pill via children to the OpportunityTypeSummary
Per design, these UI tweaks will apply to all usages of the shared component. The zero grants pill will only apply to the insights version of the component.

GitOrigin-RevId: 15e1bc403134fb39503596d72d1e0255b5d04d6e