[Workflows][US_AZ] Add subcategories of Submitted to opportunity Update Status menu (Recidiviz/recidiviz-dashboards#6574)
Allow the "Update Status" dropdown in an opportunity sidebar to potentially (1) contain multiple buttons that each put the opportunity in a different subcategory of "Submitted" and (2) vary in the possible subcategories an opportunity is eligible for. Also, begin to enforce a strict order for subcategories that will be shown both in the caseload grid display and in the "Mark Submitted" dropdown menu.

GitOrigin-RevId: 449868fe073ce79644decfd3a354be36e9b71f1d