[Workflows][Actionability] Navigate through opportunities smoothly even if statuses change (Recidiviz/recidiviz-dashboards#6763)
Maintain the list of opportunities that a user can navigate through with the opportunity preview modal footer, even if the status of those opportunities changes. Previously the list was based on the contents of the user's current tab, which broke the navigation if the currently viewed opportunity changed status (such as if the user denied or submitted it). (also make a small change to an unrelated file to address a mobx warning in the console)

GitOrigin-RevId: 3512f8d67fcfb1ce94831f52850e28dbfd657581