[Insights] Update supervisor presenter to allow for an empty list of officers (Recidiviz/recidiviz-dashboards#6842)
Updates the SupervisionSupervisorPresenter to no longer throw if the supervisor does not have outcomes officers.

The SupervisionOfficersPresenter (for the v1 supervisor page) should still throw because we would not expect a supervisor surfaced there to have no outcomes officers. Also changes the expectOfficersWithOutliersPopulated to expectOfficersWithOutcomesPopulated because it's clearer.

This is not an issue yet but we plan to start sending supervisors without outcomes officers through the backend soon.

GitOrigin-RevId: a479665539b4926bdb75fbf6162e3960c04dc0ba