[Insights] Add SupervisionOfficerOutcomes Zod schema (Recidiviz/recidiviz-dashboards#6804)
Adds a supervisionOfficerOutcomes Zod schema.

This Zod schema is made up of the outcomes-related fields from the supervisionOfficerSchema. It will be used as we start to split outcomes information from the officer schema and consume the outcomes API from the backend. Backend schema is here for reference.

PR 1/7 for this ticket. Next PR: Recidiviz/recidiviz-dashboards#6805

GitOrigin-RevId: cf5cba94b431257ab5890762b8a197f7b1cb49d3