[Insights] Use outcomes endpoint in the SupervisionSupervisorPresenter (Recidiviz/recidiviz-dashboards#6807)
Updates the SupervisionSupervisorPresenter to populate outcomes information using the new API.

This PR adds a population method to the supervision store for officer outcomes. It updates the SupervisionSupervisorPresenter to hydrate and expect that outcomes data is populated. It then modifies the outlierDataOrError fn as well as the util fn getOutlierOfficerData to use the outcomes data instead of the officer data. Adds some TODOs for further clean up once we've moved all presenters over to using the new API and modified the officer schema to remove obsolete fields.

GitOrigin-RevId: bf5a4625b46d712ad2f01dadafd8ace1de22f7fd