[Workflows][US_TN] Change opportunity status to Submitted upon downloading TEPE note (Recidiviz/recidiviz-dashboards#6858)
When a user writes to Tennessee's OMS successfully for an opportunity, which currently exists only for the TEPE note download in the supervision Expiration opportunity, automatically change the opportunity's status to the "Pending" (aka submitted/in progress) status. This change is for Actionability so the user doesn't have to both write to OMS and also mark the person as submitted. Actionability has not launched in TN, so this does not affect any state users right now.

GitOrigin-RevId: 0058f760d3be9f9a1ebad9d305327d93c289f55e