[Insights] Add outcomes API to officer presenters (Recidiviz/recidiviz-dashboards#6832)
Updates the SupervisionOfficerPresenter, SupervisionOfficerDetailPresenter, and base presenter to populate outcomes information using the new API.

This PR updates the individual officer presenters to hydrate and expect that outcomes data is populated. It modifies the outlierDataOrError fn to use the outcomes data instead of the officer data. It adds logic to the hydration/expect functions to avoid populating outcomes if the officer is excluded.

GitOrigin-RevId: 9c2c4fb00d478feae53f706ded8f6b82398bbe37