[Insights] Hide included officers tooltip heading when no included officers (Recidiviz/recidiviz-dashboards#6884)
Hides the extra heading in the "Team" tooltip for supervisors when they don't have outcomes officers

We already hide the excluded officers heading when the supervisor has no excluded officers so this is just achieving the same effect now that we want to pull in supervisors without outcomes officers. Otherwise, we would just show a header without any included officers listed.

GitOrigin-RevId: bc92c1710029da3f1d05cb8ac9f524091d0a3ba2