[Insights][Bug] Stop throwing error in highlighted officer fn (Recidiviz/recidiviz-dashboards#6933)
Fix for prod bug in the supervisor homepage officer page (excluded officers only)

After switching over the highlightOfficer function to use officer outcomes objects in its calculations, I forgot to add an isExcludedOfficer check to the getter on the officer page presenter. Because of this, we threw an error on the supervisor homepage staff page for excluded officers.

GitOrigin-RevId: e565936b7ba62a2514c8bd11af580fae4ced18c3