[Insights] Remove excluded officer logic from the getOfficerOutcomesData fn (Recidiviz/recidiviz-dashboards#6997)
Modifies the getOfficerOutcomesData fn and return type to stop handling excluded officers.

This change also stemmed from feedback on Recidiviz/recidiviz-dashboards#6834 - regarding the fact that we don't need to call this get fn for excluded officers or even have excluded officers in the return type. Moved the refactor into it's own PR though, since it was possible to do so, and it was making the original PR quite big. Also, in modifying the getOfficerOutcomesData fn to no longer be generic, the officer presenter itself no longer needs to be generic, so this PR simplifies a lot of the officer presenter code imo.

GitOrigin-RevId: 8637cecfa12dcb0dba49b436b8751d4a34b12b70