[Insights] Make avg daily population nullable on officer Zod schema (Recidiviz/recidiviz-dashboards#6882)
Makes the avgDailyPopulation field nullable on the officer schema, since some excluded officers don't have that metric calculated.

The ticket: #recidiviz-data/35315, has all the context, but to summarize, because we now are pulling in excluded officers to the product views, we don't have avg daily population calculated for some previously excluded officers. The db schema was updated already to allow for nullable metric values. Because of this, the avg_daily_population field should be nullable/optional in the FE+BE officer schema before we can fully consolidate the excluded/included officer types.

On the FE, the avgDailyPopulation is passed as an infoItem to the insights page layout. Null values just mean that that particular info item will not be rendered.

GitOrigin-RevId: a8300d6ff8b07482bbca85fb2bc61979a89e982f