[Insights] Modify excluded officer typecheck to prepare for endpoint change (Recidiviz/recidiviz-dashboards#7062)
Modifies the isExcludedOfficer type check to accommodate the changed officer model.

I discovered during testing of the BE change to send excluded officers through the /officer endpoints, that excluded officers were not passing the isExcludedOfficer type check. This modifies the type check to remove a now obsolete requirement that excluded officers cannot be parsed by the supervisionOfficerSchema. We cannot enable the BE endpoint change until this lands.

GitOrigin-RevId: 842dc6138c765814638442ecd4237fcffab6c7a2