[Insights] modify officersBySupervisorPseudoId to store all officers, excluded and included (Recidiviz/recidiviz-dashboards#7069)
Modifies the officers to Supervisor mapping in the InsightsSupervisionStore to take advantage of new insights endpoint behavior (all officers, included and excluded, are now coming through the /officers endpoint)

This is PR 1/5 for Recidiviz/recidiviz-dashboards#6453, which encapsulates the work to remove the insights FE dependency on the /excluded_officer endpoints and ExcludedOfficer zod schema. Next PR is Recidiviz/recidiviz-dashboards#7070.

GitOrigin-RevId: a1b9651157157c5d77f23bd00e17c95d206b39db