[Insights] Stop using excludedOfficersBySupervisorPseudoId mapping in vitals and supervisor presenter (Recidiviz/recidiviz-dashboards#7070)
This is PR 2/5 for Recidiviz/recidiviz-dashboards#6453, which encapsulates the work to remove the insights FE dependency on the /excluded_officer endpoints and ExcludedOfficer zod schema.

This PR also makes the fixture update to add two officers to the supervisionOfficerFixture where includeInOutcomes = false (moves them from the excludedSupervisionOfficerFixture which will soon be deleted). Relevant snapshots are updated.

GitOrigin-RevId: 659cc09e128c2e12fafa09e4033e55469cd5fa09