[Insights] Remove ExcludedOfficer type and endpoint usage from the InsightsSupervisionStore (Recidiviz/recidiviz-dashboards#7071)
Removes the ExcludedSupervisionOfficer type from the insightsSupervisionStore, as we no longer need it given the consolidate Officer type and API.

This is PR 3/5 for Recidiviz/recidiviz-dashboards#6453, which encapsulates the work to remove the insights FE dependency on the /excluded_officer endpoints and ExcludedOfficer zod schema. Prior PR is Recidiviz/recidiviz-dashboards#7070. Next PR is Recidiviz/recidiviz-dashboards#7072.

GitOrigin-RevId: cd212b4828c53caea3d49483b2b560ec92e6f82b