[Insights] Remove excluded officer endpoint usage in SupervisionOfficerPresenter (Recidiviz/recidiviz-dashboards#7072)
In the individual officer presenter, remove the call to the /excludedOfficer endpoint - it's no longer necessary with all officers being surfaced by the /officer endpoint. Also remove some union types.

This is PR 4/5 for Recidiviz/recidiviz-dashboards#6453, which encapsulates the work to remove the insights FE dependency on the /excluded_officer endpoints and ExcludedOfficer zod schema. Prior PR is Recidiviz/recidiviz-dashboards#7071. Next PR is Recidiviz/recidiviz-dashboards#7073.

GitOrigin-RevId: ffedac633ddfb00bdb42e3035e11501d733547b1