[Insights] Remove excluded officer type from the isExcludedSupervisionOfficer fn (Recidiviz/recidiviz-dashboards#7073)
Officer types have been consolidated, so rather than type narrowing, the isExcludedSupervisionOfficer function returns a boolean.

This is PR 5/5 for Recidiviz/recidiviz-dashboards#6453, which encapsulates the work to remove the insights FE dependency on the /excluded_officer endpoints and ExcludedOfficer zod schema. Prior PR is Recidiviz/recidiviz-dashboards#7072.

There are a couple more clean up todos related to this ticket, but they are negligible as far as functionality, so this PR ends the stack!

GitOrigin-RevId: afccfd5d9158f9824fc6d6d131b74c3c258c006a