[Workflows] Only set last viewed for visible opportunities on profile page (Recidiviz/recidiviz-dashboards#7112)
Only set an opportunity as "last viewed" in Firestore if that opportunity is expanded within an accordion, and track the opportunity as having been previewed, for more accurate impact analytics. The opportunities accordion shows up on the JII profile page and in the sidebar in Tasks. Previously, opportunities were set as "last viewed" if a JII's profile page was viewed even if the corresponding accordion item hadn't been opened.

GitOrigin-RevId: 4157e59d4de29e70c6fc06629db55d84c8883104