[Workflows] Mark opportunities as Submitted when form is downloaded (Recidiviz/recidiviz-dashboards#7107)
Automatically change an opportunity's status to Submitted (and display a toast) whenever its form is downloaded. This is gated behind the partially rolled out submittedOpportunityStatus feature variant. For the edge case of Arizona (currently the only state where we might have multiple "subcategories" of the "submitted" status), we will NOT automatically change the opportunity's status upon form download - the user still has to select a subcategory themself. This PR also adds some checks to make sure we don't mark an opportunity as submitted when we are in that case, as well as error throwing if we somehow get a bad request for marking an opportunity as submitted.

GitOrigin-RevId: 49012d884e3ddd1f666a3527d619bc8dccd7808c