[Hydration Utils] add optional loading prop (Recidiviz/recidiviz-dashboards#7237)
* feat(Hydrator): add optional loading prop

The changes add an optional `loading` prop to the `Hydrator` component in the `hydration-utils` library. This prop allows the caller to provide a custom loading state to be displayed when the hydratable item is in the "loading" state.

The `ModelHydrator` component in the `staff` app has also been updated to pass the `loading` prop to the `Hydrator` component, allowing the app to customize the loading state.

The `Hydrator` component has been updated to use the `loading` prop if it is provided, otherwise, it will display the default loading state.

* feat(hydration-utils): handle null loading state in Hydrator

Adds a new case to handle the `null` loading state in the Hydrator component.
This ensures that the component correctly handles the case where the loading
state is explicitly set to `null`, which indicates that no loading state
should be displayed.

GitOrigin-RevId: 421744654858f9856726468feccd6b6b2aeaac83