feat(insights-page): Add methodology link and cta (Recidiviz/recidiviz-dashboards#7272)
This change adds a methodology link and call-to-action (CTA) to the
Insights Page Section component. The methodology link and CTA are
displayed alongside the section description, allowing users to access
more detailed information about the insights.

The key changes are:

- Add `methodologyLink` and `methodologyLinkCta` props to the
  `InsightsPageSectionProps` type
- Conditionally render the methodology link and CTA in the
  `InsightsPageSection` component
- Style the methodology link using the `StyledLink` component

These changes improve the user experience by providing a clear path for
users to access more information about the insights presented on the
page.

GitOrigin-RevId: e47e326568d806ad85f4775c014eb1a6cbc0674b