[Workflows][US_AZ] Add CSED date to resident metadata and dates table (Recidiviz/recidiviz-dashboards#7321)
Add the CSED (Community Supervision End Date) to the dates table displayed in the resident profile and sidebar. Testers requested visibility of this date and we believe it'll help with tool adoption. The CSED is calculated based on 115% of the sentence time, so it always comes after the SED (Sentence End Date = 100% of the sentence).

GitOrigin-RevId: 2e83ad703c908a4f870f53349b961f111180b6a3