[Workflows][US_UT] Create Early Termination opportunity class, demo mode fixtures  (Recidiviz/recidiviz-dashboards#7508)
Create a base opportunity class for the Utah Early Termination of Probation opportunity. Also create demo mode fixtures for supervision staff, clients, and referrals for this opportunity. Everything is as minimal as possible right now - I will fill in the referrals and mock config for this opportunity more once data is flowing and we have the copy.

GitOrigin-RevId: 2bf946e660880b308b72eeb796d3612d9b5f9e11