Add deployer's name and list deployed services in Slack bot post (Recidiviz/recidiviz-dashboards#7627)
Add some info to the Polaris staging deploy message posted by the Deployment Bot slack account: Add the email of the person who did the deploy and a list of deployed services to the Polaris deploy notification posted in Slack, so that it's more obvious when someone has done a partial deploy. Also for staging deploys, specify when one is being done dangerously, i.e. not from the latest commit on main.

GitOrigin-RevId: 5767bc0ff4065abca004e3940b3a01644d366183