[Insights] Show active clients count on officer metric page (Recidiviz/recidiviz-dashboards#7656)
Adds the current client count to the officer metric page.

Adds populateCaseloadForOfficer to the JIIMixin class so that we can populate the client caseload map for the given officer without hydrating all client opportunities.
Modifies the metric page's presenter to use the mixin.
Adds an info item for the current client count (named "active clients" to align w the officer page).

GitOrigin-RevId: 5afaabca98c1593f4b41ba3009bdfb8812c7c1a9