[Insights][RIR] Implement `submitRosterChangeRequestIntercomTicket` and `populateAllSupervisionOfficers` for the `InsightsSupervisionStore` (Recidiviz/recidiviz-dashboards#7540)
* snap fail

* feat(analytics): Add new analytics tracking for roster modal and change request

The changes in this commit add new analytics tracking for the following events:

1. Roster modal viewed
2. Roster change request form submitted

These new tracking events will help us better understand user behavior and interactions with the roster functionality.

* add id to analytics

* feat(supervision): add ability to fetch all supervision officers

This change adds the ability to fetch all supervision officers from the
Insights API. This is necessary to support the new roster change request
feature, which requires access to the full list of supervision officers.

The new `populateAllSupervisionOfficers` method is added to the
`InsightsSupervisionStore` to handle the API call and store the results.
Additionally, a new `supervisionOfficers` getter is added to provide access
to the fetched officers.

* feat: add "reportIncorrectRosters" feature variant

The changes add a new feature variant called "reportIncorrectRosters" to the
application. This feature variant is used to determine whether a user can
submit a roster change request. The changes include:

- Adding the "reportIncorrectRosters" feature variant to the
  `FeatureVariant` type and the `allFeatureVariants` mapping in the
  `RootStore/types.ts` file.
- Updating the `userCanSubmitRosterChangeRequest` getter in the
  `InsightsSupervisionStore` to check for the presence of the
  "reportIncorrectRosters" feature variant.
- Adding tests to ensure the `userCanSubmitRosterChangeRequest` getter
  behaves as expected when the "reportIncorrectRosters" feature variant
  is present or absent.

* Gina's comments

* code change

* Dana's comments

* update tests

GitOrigin-RevId: e8e092d532d154cb10cce838c91b27ccf968d282