[Insights] Add email to SupervisionOfficer Zod schema and display info item (Recidiviz/recidiviz-dashboards#7698)
Adds officer email as a nullable field on the officer zod schema and displays the email on the officer overview page.

Final PR to add missing info items to supervisor homepage. We had previously (see #data/39521) started sending the officer email from the BE so this PR updates the zod schema and the officer page to use the field.

Updated snapshots in staff and datatypes. Added fixtures including one with a null value. Preview URL looks good!

Backend change goes to prod on Tuesday Mar 18 - will wait to land this til after the prod deploy.

GitOrigin-RevId: c33055651c0ce6a7ab920d60207284caecaa3048