[Workflows][US_UT] Enable feature variants for Recidiviz users in Utah in prod (Recidiviz/recidiviz-dashboards#7779)
Add these FVs for Recidiviz users in Utah in prod: fullWidthTimeline, nonOMSCriteria, oppTabSubcategories, sortableOpportunityTabs, submittedOpportunityStatus

This is the same as the current set of default permissions for supervision line staff in Utah, so enables Recidiviz users to see what real users see in prod. This is also all of the feature variants that are applicable to Workflows and not launched to all states, although maybe fullWidthTimeline should be ungated at this point.

GitOrigin-RevId: 3c025c8f0665adc5fa297d6fdb6f66b81d821f60