[Workflows] Add release & sentence end dates to Table View (Recidiviz/recidiviz-dashboards#7924)
Add new columns to Workflows Table View that display a person's release date from incarceration and the end date of their supervision sentence. The appropriate column is shown depending on whether we're currently showing facilities or supervision opportunities. We'd like to enable users to view and sort by this date when it is applicable to the opportunity - most opportunity types already silently sort by release date.

GitOrigin-RevId: 3833cff570aaa26d7ee4deaf5b1d932219110b7f