[Workflows] Show all tab names in opportunity CTA (Recidiviz/recidiviz-dashboards#7953)
Show "eligible now" and "almost eligible" on the opportunity CTA elements used in the Workflows homepage and Supervisor Homepage Opportunities Module, to reduce user confusion. To make this general, this PR will show the non-empty tab names that apply to a certain opportunity.

Also, make the number shown next to the avatars on the right hand side, if there are more than 5 people eligible for an opportunity, match the number shown in the call to action.

GitOrigin-RevId: 5baef1921ffbb3a2c920daabfb8f96207d57e396