Accept LSU records without UA criteria (Recidiviz/recidiviz-dashboards#7977)
Recidiviz/recidiviz-dataRecidiviz/recidiviz-dashboards#40587 is removing the urinalysis/drug test/UA criteria (negativeDAWithin90Days) for the LSU opportunity in Idaho. This may not be a permanent change - there will be some validation calls next week. This PR makes the UA criteria optional in the LSU zod schema and where it's checked in the opportunity class, so that we won't get errors when we start receiving opportunity referrals in staging that lack this criteria item.

GitOrigin-RevId: 8a1c750fb43d7d1c6375a42a33a0ce43d0f36937