[Workflows][US_ID] Clean up old CRC zod schemas (Recidiviz/recidiviz-dashboards#7837)
Follow-up to Recidiviz/recidiviz-dashboards#7808.

Recidiviz/recidiviz-dataRecidiviz/recidiviz-dashboards#38299 is changing the shape of CRC opportunity records. After that change goes to prod, the zod schema for parsing/validating the old shape will no longer be needed, and having it in the codebase will slow down the parsing of opportunity records (because zod tries to parse union members in order).

GitOrigin-RevId: 68d0c3fe35d4185ac749fc2b1a6ff9c19bfa4cdc