[Workflows][US_TN] Use new supervision level copy mappings for SLD (Recidiviz/recidiviz-dashboards#8074)
This change relaxes the TN SLD validator to surface opportunities even if the supervision level in the criteria doesn't match the supervision level on the client record. We need to do this because our internal supervision level label used in criteria ("medium") doesn't match either the TN raw text ("8MO") or external copy ("moderate") used in the client record.

GitOrigin-RevId: d58d2a6470a8ea4519c9ecc923bfc6da7f12fc7e