[Workflows][US_AZ] Make eligibility date labels more precise for DTP/TPR (Recidiviz/recidiviz-dashboards#8127)
This PR introduces an eligibilityDateTextForTab method on the opportunity configuration. The AZ DTP/TPR custom configs override this to show the right copy for each tab.

For AZ DTP/TPR opportunities, we treat the DTP/TPR date as the eligibility date for Table View. For some people, we pull this date from AZ's OMS, while for others, we calculate a projected date ourselves. We would like to display only "DTP Date" where appropriate and display "DTP or Projected DTP" otherwise, and the same for TPR. However, previously we had only a string field (eligibilityDateText) for the label for eligibility date, so we couldn't customize this by tab.

GitOrigin-RevId: c7ee90a21d535207aea091185e91d90e537bbd22