[Workflows][US_TN] Disable old Compliant Reporting for users with 2025 FV (Recidiviz/recidiviz-dashboards#8175)
Starting on 5/1, TN will be rolling out a new compliant reporting policy to certain users in some pilot districts. We want those users to only see the new "2025 policy" CR opportunity and not the old CR opportunity, and all other users to only see the old opportunity and not the new one. To do so, the new opportunity is gated behind the usTnCompliantReporting2025Policy feature variant, which Damini will be giving to the pilot users.

This PR disables the old Compliant Reporting opportunity for anyone with the 2025 policy FV by using the "inverse feature variant" config feature. This will have no effect for anyone (except for Recidiviz users) until we assign the FVs to real users.

GitOrigin-RevId: 4ba82ea121c62e865e95e845f974af09321dd45c