[Tasks] Switch Tasks gating to specific tasks route, part 1/2 (Recidiviz/recidiviz-dashboards#8163)
This is half of a change that will allow Workflows users to see tasks without being able to see opportunities/workflows. We wanted this to happen for Texas users before I realized the second half of the change was a little complicated right now and we didn't absolutely need to do that. So mainly this change renames the relevant WorkflowsStore properties to be a little clearer.

GitOrigin-RevId: 0f8bac80825eb02f5bed7cb4c98fe292fceaf0fb