[Insights][RIR][Bug[ Add selected indicator to officer select UI for roster change request modal dropdown (Recidiviz/recidiviz-dashboards#8077)
* 🎨 feat(staff): enhance supervision officer select UI

Improve the UI of the supervision officer select dropdown in the
Insights Roster Change Request form. Add a `FlexWrapper` component to
display the request change type option and a `SelectedOpportunityTypeIndicator`
when the option is selected. This provides a clearer visual indication
of the selected option.

* 🎨 feat(staff): improve opportunity type select UI

This commit introduces the following improvements to the opportunity type select UI:

- Replaces the custom selected opportunity type indicator with an icon
- Removes the disabled state from the dropdown menu items, as it was causing confusion
- Renames the `disabled` prop to `selected` to better reflect its purpose

These changes aim to enhance the user experience and make the UI more intuitive.

* rename SelectedOpportunityTypeIndicator to SelectedCheckmarkIndicator

* change width

GitOrigin-RevId: 1e3b90f9179876831233810f09a8f90d83857f25