[Workflows] Add supervision end date copy override, AR release date copy (Recidiviz/recidiviz-dashboards#8066)
Create a new tenant config field supervisionEndCopyOverride , similar to releaseDateCopyOverride but for supervision end dates. We want these to consistently be labeled as "expiration" dates in Tennessee.

Also add a release date copy override for Arkansas, and put the state-specific copy in the Marked Ineligible reasons where we had previously used generic copy. The latter only appears in the edge case that the person's snooze end date is the same as the person's expiration or release date

GitOrigin-RevId: 27c4deb413975b24d71e045651651532485adfca